Replace ULogDataProvider with UlogIterableSource
Fixes: #3979
